requirements.txt
fastapi 
uvicorn 
sqlalchemy 
mysqlclient 
pymysql 
pydantic
psycopg2-binary
python-dotenv
